% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.read_dstrace_subset}
\alias{.read_dstrace_subset}
\title{Load a cropped version of downscaled TraCE21ka .nc files}
\usage{
.read_dstrace_subset(file, sf = NULL)
}
\arguments{
\item{file}{A string with full name (path and file name) of the .nc file to be loaded from the hard disk drive.}

\item{sf}{A sf object from the sf package to be used as template or mask to crop the downscaled trace .nc file.}
}
\value{
A stars object with the downscaled trace data.
}
\description{
Load a cropped version of downscaled TraCE21ka .nc files
}
\examples{
# TBW
}
