% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_dscmip_subset}
\alias{read_dscmip_subset}
\title{Read downscaled CMIP5 datasets}
\usage{
read_dscmip_subset(folder, var, scen, gcm, sf = NULL)
}
\arguments{
\item{folder}{TBW}

\item{var}{TBW}

\item{scen}{TBW}

\item{gcm}{TBW}

\item{sf}{TBW}
}
\value{
TBW
}
\description{
Read downscaled CMIP5 datasets
}
\examples{
#TBW  
}
