% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_trace_subset}
\alias{read_trace_subset}
\title{Read original TraCE21ka datasets}
\usage{
read_trace_subset(folder, var, point)
}
\arguments{
\item{folder}{TBW}

\item{var}{TBW}

\item{point}{TBW}
}
\value{
TBW
}
\description{
Read original TraCE21ka datasets
}
\examples{
# TBW 
}
