% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_trace_subset}
\alias{read_trace_subset}
\title{Load downscaled TraCE21ka files by years.}
\usage{
read_trace_subset(folder, var, y_start, y_end, sf = NULL)
}
\arguments{
\item{folder}{A character string with the path to the folder where the downscaled TraCE21ka files, ordered in folders by variables.}

\item{var}{A character string with the name of the variables to be loaded.}

\item{y_start}{A number with the first year to be loaded.}

\item{y_end}{A number with the last year to be loaded.}

\item{sf}{A sf object to be used for cropping the object.}
}
\value{
A stars object with the downscaled TraCE21ka data, cropped using the sf object if provided.
}
\description{
Load downscaled TraCE21ka files by years.
}
\details{
Years should be in the format calibrated Before Present (e.g. 0 for loading year 1950 in the gregorian calendar).
}
\examples{
# TBW
}
